requirements.txt
streamlit==1.36.0
anthropic==0.45.2
azure-ai-inference==1.0.0b9
azure-cognitiveservices-speech==1.31.0
python-dotenv==1.0.1
aiohttp==3.11.11
httpx==0.28.1
numpy==1.26.4
pandas==2.2.1
fastapi==0.115.7
uvicorn==0.29.0
pyyaml==6.0.1
packaging==24.0
tiktoken==0.6.0 